requirements.txt
fyers-apiv3
pandas
numpy
matplotlib
seaborn
scikit-learn
statsmodels
python-dotenv
